Data.Compact.Incremental

Description

This module provides a data structure, called a Compact, for holding a set of fully evaluated Haskell values in a consecutive block of memory.

As the data fully evaluated and pure (thus immutable), it maintains the invariant that no memory reference exist from objects inside the Compact to objects outside, thus allowing very fast garbage collection (at the expense of increased memory usage, because the entire set of object is kept alive if any object is alive).

This is similar to Data.Compact, but it supports incremental evaluation into Compact structures.

Since: 1.0.0

Documentation

data Compact a

A pointer to an object that lives in a compact region.

compactGetRoot :: Compact a -> a

Returns the underlying object pointer of a compact pointer.

compactResize :: Compact a -> Word -> IO ()

Resizes a compact region to a new size. Existing data is preserved.

compactContains :: Compact b -> a -> Bool

Tests if a lives in the compact region Compact b.

compactContainsAny :: a -> Bool

Tests if a is in compact normal form (i.e. is in any compact region)

compactNew :: Compactable a => Word -> a -> IO (Compact a)

compactNewAt :: Compactable a => Word -> Ptr b -> a -> IO (Compact a)

compactNewSmall :: a -> IO (Compact a)

Creates a "fake" compact region suitable for storing very small objects, where actually trying to make them contiguous is not profitable.

compactAppendOne :: Compactable a => Compact b -> a -> IO (Compact a)

compactAppendRecursively :: Compactable a => Compact b -> a -> IO (Compact a)

compactAppendEvaled :: Compact b -> a -> IO (Compact a)

compactAppendSmall :: Compactable a => Compact b -> a -> IO (Compact a)

class Compactable a where

Methods

compact :: Compact b -> a -> IO (Compact a)

defaultCompactNFData :: NFData a => Compact b -> a -> IO (Compact a)

defaultCompactNFData: a default implementation for compact suitable | for NFData instances

data SerializedCompact a

A metadata structure which represents a compact region as a (list) of pointers to memory regions (with their lengths), with a distinguished pointer representing the root of the region. This structure is transmitted out of band along with the actual contents of the regions.

Constructors

SerializedCompact

Fields

serializedCompactGetBlockList :: [(Ptr a, Word)]

serializedCompactGetRoot :: Ptr a

withCompactPtrs :: (Compactable a, NFData c) => Compact a -> (SerializedCompact a -> IO c) -> IO c

compactImport :: SerializedCompact a -> (Ptr b -> Word -> IO ()) -> IO (Maybe (Compact a))

Given a SerializedCompact data structure, as well as a "fill function" which loads data to the address 'Ptr b', import a compact region into Haskell. We're parametrized over a fill function so that a user either copy or mmap memory into the correct position (the user does not generally know what an acceptable address to load the data structure is.) Note that the values passed by 'Ptr b' may vary from the pointers specified in SerializedCompact; if this is the case, compactImport will fix up the pointers so they point to the correct location.

compactImportTrusted :: SerializedCompact a -> (Ptr b -> Word -> IO ()) -> IO (Maybe (Compact a))

Variant of compactImport that assumes that the data comes from a source we trust. This runs faster but is less safe.

compactImportByteStrings :: SerializedCompact a -> [ByteString] -> IO (Maybe (Compact a))

Easier to use version of compactImport which works on Bytestrings.

compactImportByteStringsTrusted :: SerializedCompact a -> [ByteString] -> IO (Maybe (Compact a))

Easier to use version of compactImport which works on Bytestrings and assumes the input is trusted.

compactInitForSymbols :: Compact a -> IO ()

This enables symbol table support for compact regions, which means that the compact region can be sent to another binary which is not identical to our source binary. You must still call compactBuildSymbolTable before sending.

compactBuildSymbolTable :: Compact a -> IO ()

Call this prior to transmitting a compact region; this creates a symbol table which can be used to fix up info tables at the receiving end. (This symbol table is automatically transmitted with the rest of the compact region.)