Data.Compact.Monad

Description

This module provides a data structure, called a Compact, for holding a set of fully evaluated Haskell values in a consecutive block of memory.

As the data fully evaluated and pure (thus immutable), it maintains the invariant that no memory reference exist from objects inside the Compact to objects outside, thus allowing very fast garbage collection (at the expense of increased memory usage, because the entire set of object is kept alive if any object is alive).

This module provides a Monad for executing code in the context of a compact, so that all values threaded through the monad are stored in the compact.

Since: 1.0.0

Documentation

data CompactM b a

runCompactM :: Compactable b => CompactM b a -> Compact b -> IO (Compact a)

compactPut :: Compactable a => a -> CompactM b a